use older prefetch-dependencies-oci-ta bundle
The newer bundle is not downloading required .tar.gz files.
https://redhat-internal.slack.com/archives/C04PZ7H0VA8/p1765203227359139

Downgrade to older bundle in the interim to allow for the pipeline to build
for 3.2. The better/long-term solution is being discussed.

odh-feature-server-v3-2 seems to be affected and needs this
task bundle to be downgraded to be able to build in Konflux.